pyproject.toml
[tool.poetry]
name = "board-game-scraper"
version = "3.0.0-alpha.1"
description = "Scraping data about board games from the web"
authors = [
    "Markus Shepherd <markus@recommend.games>",
]
license = "MIT"
readme = "README.md"

documentation = "https://recommend-games.github.io/board-game-scraper"
homepage = "https://recommend-games.github.io/board-game-scraper"
repository = "https://github.com/recommend-games/board-game-scraper"

classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Operating System :: OS Independent",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Typing :: Typed",
]

packages = [
    { include = "board_game_scraper", from = "src" }
]

[tool.poetry.dependencies]
python = ">=3.13.0, <3.14.0"

attrs = "*"
itemadapter = "*"
itemloaders = ">=1.1.0, <2.0"
jmespath = "*"
more-itertools = "*"
parsel = ">=1.8.1, <2.0"
python-dateutil = "*"
pyyaml = "*"
requests = "*"
scrapy = ">=2.13.0, <3.0"
scrapy-extensions = {version = ">=1.1.0", extras = ["blurhash"]}
w3lib = ">=2.0.0, <3.0"

[tool.poetry.group.dev.dependencies]
ipython = "*"
mkdocstrings = {version = ">=0.23", extras = ["python"]}
mkdocs-material = "*"
mypy = "*"
pre-commit = "*"
pymdown-extensions = "*"
pytest = "*"
pytest-github-actions-annotate-failures = "*"
pytest-cov = "*"
python-dotenv = {version = "*", extras = ["cli"]}
python-kacl = "*"
ruff = ">=0.2.0"
types-jmespath = "*"
types-python-dateutil = "*"
types-pyyaml = "*"
types-requests = "*"

[tool.poetry.scripts]
bg-scraper = "board_game_scraper.__main__:main"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
target-version = "py313" # The lowest supported version

[tool.ruff.lint]
# By default, enable all the lint rules.
# Add to the ignore list below if you don't want some rules.
# If you need some ignores for certain modules, see tool.ruff.lint.per-file-ignores below.
# For individual ignore cases, prefer inline `# noqa`s within the code.
select = ["ALL"]
ignore = [
    "ANN",      # Type hints related, let mypy handle these.
    "D",        # Docstrings related, way too strict to our taste
    "ERA001",   # Found commented-out code
    "RUF001",   # String contains ambiguous `–` (EN DASH)
    "TD",       # TODO
    "FIX",      # FIXME,
    "S101",     # "Use of `assert` detected"
    "T201",     # Removes `print`s – WTF?
    "PLC0415",  # `import` should be at the top-level of a file
]

[tool.ruff.lint.per-file-ignores]
"tests/**" = [
    "S101",     # "Use of `assert` detected"
    "ARG",      # "Unused function argument". Fixtures are often unused.
    "S105",     # "Possible hardcoded password".
]

[tool.ruff.lint.mccabe]
max-complexity = 20

[tool.ruff.lint.pep8-naming]
classmethod-decorators = [
    "classmethod",
    "pydantic.validator",
    "pydantic.root_validator",
]

[tool.ruff.lint.pylint]
max-args = 10
max-statements = 75

[tool.pytest.ini_options]
addopts = """\
    --cov board_game_scraper \
    --cov tests \
    --cov-report term-missing \
    --no-cov-on-fail \
"""

[tool.coverage.report]
fail_under = 100
exclude_lines = [
    'if TYPE_CHECKING:',
    'pragma: no cover'
]

[tool.mypy]
# This is the global mypy configuration.
# Avoid changing this!
strict = true  # See all the enabled flags `mypy --help | grep -A 10 'Strict mode'`
disallow_any_unimported = false
disallow_subclassing_any = false

# If you need to ignore something for some specific module,
# add overrides for them. Avoid changing the global config!
# For example:
[[tool.mypy.overrides]]
module = [
    "itemadapter.*",
    "scrapy.*",
]
ignore_missing_imports = true
implicit_reexport = true

# [[tool.mypy.overrides]]
# module = [
#   "tests/my_thing/test_my_thing",
# ]
# disallow_untyped_defs = false
